feat: put full message into main section
