+++ src/components/ButtonLink/ButtonLink.module.scss
@@ -1,9 +1,14 @@
 .button_home {
+  box-sizing: border-box;
+  position: absolute;
+  top: 0;
+  right: 0;
+
   border-radius: 2px;
   border: 1px solid transparent;
   box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.24), 0 0 1px 0 rgba(0, 0, 0, 0.24);
   color: rgba(0, 0, 0, 0.54);
-  margin: 1rem 0;
+  margin: 1rem ;
   height: 4rem;
   width: 10rem;
   background: white;
