+++ src/components/Message/Message.module.scss
@@ -1,5 +1,4 @@
 .message_section_container {
-  position: absolute;
   width: 100%;
   padding-top: 3rem;
   display: flex;
@@ -9,11 +8,12 @@
 }
 
 .message_section {
+  position: relative;
   display: flex;
   flex-direction: column;
   padding: 1rem;
-  width: 90%;
-  box-shadow: 0 14px 28px rgba(0, 0, 0, 0.25), 0 10px 10px rgba(0, 0, 0, 0.22);
+  width: 100%;
+  background: #fcfdfe;
   overflow-wrap: break-word;
 }
