+++ src/pages/MessagePage/MessagePage.tsx
@@ -1,11 +1,12 @@
 import React, { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import { loadOneMessage } from '../../api/userMessages/userMessages';
-import { ButtonLink } from '../../components/ButtonLink/ButtonLink';
+// import { ButtonLink } from '../../components/ButtonLink/ButtonLink';
 import { Message } from '../../components/Message/Message';
 import { GoogleMessage } from '../../models/GoogleMessage';
 import { sortMessageData } from '../../hooks/useSortMessageData/useSortMessageData';
 import base64url from 'base64url';
+import { Main } from '../../components/Main/Main';
 
 interface MessagePageValues {
   snippet?: string;
@@ -17,7 +18,7 @@ interface MessagePageValues {
 
 export const MessagePage: React.FC = () => {
   const [message, setMessage] = useState<GoogleMessage | undefined>();
-  const { search, state } = useLocation();
+  const { state } = useLocation();
   const apiData: any = state;
   useEffect(() => {
     loadOneMessage(apiData.id, apiData.token).then((message) =>
@@ -39,18 +40,19 @@ export const MessagePage: React.FC = () => {
   let convertedSuperText = '';
   if (messageData.text)
     convertedSuperText = base64url.toBase64(messageData.text);
-  const realHref = {
-    pathname: '/home',
-    search,
-  };
+  // const realHref = {
+  //   pathname: '/home',
+  //   search,
+  // };
+
   return (
-    <>
-      <ButtonLink href={realHref}>Return Home</ButtonLink>
+    <Main>
+      {/*<ButtonLink href={realHref}>Return</ButtonLink>*/}
       <Message
         from={messageData.from}
         title={messageData.title}
         text={convertedSuperText}
       />
-    </>
+    </Main>
   );
 };
